Adding identical titles from Sserpdrow Client
This commit is a near identical copy-paste of the current client side
README.md
As stated in a previous commit, moving forward we will have identical
README.md for both API and Client.
Neither is complete at this time, several more changes to come. This
will be pushed as a manner of testing that I (Sean) pulled correctly.